requirements.txt
langchain
langchain-groq
gradio
python-dotenv
pydantic
pydantic-settings
ruff
black